CORE-8643 Move DeployedComponents properties to model package
