+++ de-lib/src/main/java/org/iplantc/de/apps/integration/client/view/tools/DeployedComponentsListingViewImpl.java
@@ -4,6 +4,7 @@
 package org.iplantc.de.apps.integration.client.view.tools;
 
 import org.iplantc.de.apps.integration.client.events.ShowToolInfoEvent;
+import org.iplantc.de.apps.integration.client.model.DeployedComponentProperties;
 import org.iplantc.de.apps.integration.client.view.deployedComponents.cells.DCNameHyperlinkCell;
 import org.iplantc.de.apps.integration.client.dialogs.ToolInfoDialog;
 import org.iplantc.de.apps.integration.shared.AppIntegrationModule;
